Move graphql stuff out
